chore: refactor MachineTypeSelection / phase 2
move util
